% Auto-generated: do not edit by hand
\name{feetAnimation}

\alias{feetAnimation}

\title{FeetAnimation component}

\description{
Custom component for displaying sensors position on the feet and their current value
}

\usage{
feetAnimation(id=NULL, className=NULL, width=NULL, height=NULL, sesnorValues=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. CSS classes added to the main div}

\item{width}{Numeric. Width of the component in px}

\item{height}{Numeric. Height of the component in px}

\item{sesnorValues}{List of numerics. Feet pressure sensor values}
}
